All vehicles: increase throttle/scaled-RC resolution to float
Fork commit 5534682062 (originally branch manual_res). Promotes the
RC_Channel scaled-input pipeline from int16 to float so the VFR_HUD
throttle and OSD throttle elements expose sub-percent precision in
manual modes instead of rounding to integer percent.

RC_Channel base (libraries/RC_Channel):
- control_in member: int16_t -> float
- get_control_in() / set_control_in(): int16_t -> float
- get_control_in_zero_dz(): int16_t -> float
- pwm_to_range() / pwm_to_range_dz(): int16_t -> float, with the
  inner int32 multiply replaced by float multiply

GCS_MAVLink (libraries/GCS_MAVLink/GCS.h):
- virtual vfr_hud_throttle(): int16_t -> float
- get_hud_throttle() wrapper: int16_t -> float

Per-vehicle vfr_hud_throttle() overrides updated to match the new
virtual signature (Plane/Copter/Rover/Sub/Blimp). Tracker has no
override and uses the default return-0 from the base, untouched.

Plane:
- throttle_percentage() int8 -> float (Plane.h + system.cpp);
  constrain_int16 -> constrain_float so the precision survives.
- Attitude / mode_cruise / qautotune: replace
  channel_*->get_control_in() == 0 / != 0 with is_zero() /
  !is_zero() since the comparison source is now float.

Copter:
- RC_Channel.cpp SAVE_TRIM / SAVE_WP guards: is_zero substitution.
- compassmot.cpp throttle-not-zero check: is_zero substitution.
- GCS_Mavlink.cpp: drop the redundant (int16_t) cast on
  motors->get_throttle() * 100.

Sub / Blimp: same drop-the-cast in vfr_hud_throttle().
Blimp/RC_Channel.cpp SAVE_TRIM guard: is_zero substitution.

Rover:
- Log_Steering / Log_Throttle steering_in / throttle_in:
  int16_t -> float; logger format strings updated from h -> f
  for the affected columns.

AP_Arming:
- rc_arm_checks: get_control_in() != 0 -> !is_zero() in both
  the generic-channel loop and the throttle-channel branch.

AP_OSD:
- draw_throttle: switch %3d to %3ld + lrintf() so the integer-percent
  display rounds the float rather than truncating it.

AP_Scripting:
- bindings.desc and docs.lua updated so gcs:get_hud_throttle()
  returns a number (float) instead of int16_t; Lua callers that
  already wrap with math.floor() are unaffected.

+++ ArduPlane/GCS_Mavlink.h
@@ -85,7 +85,7 @@ private:
     MAV_STATE vehicle_system_status() const override;
 
     float vfr_hud_airspeed() const override;
-    int16_t vfr_hud_throttle() const override;
+    float vfr_hud_throttle() const override;
     float vfr_hud_climbrate() const override;
     
 #if HAL_HIGH_LATENCY2_ENABLED
